reafactor: Создает новый презентер для отрисовки точки путешествия и переносит всю логику построения точки туда
